fix: 퀴즈 토큰 저장소 파일 안티매크로 -> lib/utils로 이동
