--- src/features/anti-macro/utils/quiz-passed-token.ts
@@ -1,7 +1,0 @@
-const KEY = 'quiz_passed_token';
-
-export const quizPassedTokenStorage = {
-  save: (token: string) => sessionStorage.setItem(KEY, token),
-  get: () => sessionStorage.getItem(KEY),
-  remove: () => sessionStorage.removeItem(KEY),
-};
