+++ src/lib/utils.ts
@@ -39,3 +39,11 @@ export function formatOpenAt(openAt: string) {
   }).format(date);
   return { datePart, timePart };
 }
+
+const KEY = 'quiz_passed_token';
+
+export const quizPassedTokenStorage = {
+  save: (token: string) => sessionStorage.setItem(KEY, token),
+  get: () => sessionStorage.getItem(KEY),
+  remove: () => sessionStorage.removeItem(KEY),
+};
